requirements.txt
asgiref==3.9.1
confluent-kafka==2.12.1
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
django-phonenumber-field==8.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.5.3
gunicorn==23.0.0
inflection==0.5.1
kafka-python==2.2.15
packaging==25.0
phonenumbers==9.0.17
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.9.0
